Fix compilation in GCC 4.6

+++ src/config/opencl/common.h
@@ -86,14 +86,16 @@ public:
     inline bool operator==(const _opencl_ptr_t &ptr) const {
         return base_ == ptr.base_ && offset_ == ptr.offset_ && pasId_ == ptr.pasId_;
     }
-    inline bool operator==(int i) const {
+
+    inline bool operator==(long i) const {
         return base_ == cl_mem(i);
     }
 
     inline bool operator!=(const _opencl_ptr_t &ptr) const {
         return base_ != ptr.base_ || offset_ != ptr.offset_ || pasId_ != ptr.pasId_;
     }
-    inline bool operator!=(int i) const {
+
+    inline bool operator!=(long i) const {
         return base_ != cl_mem(i);
     }
